breaking refactor: use much lighter parse+compile engine
This package forked the very lightweight [simple-markdown](https://github.com/Khan/simple-markdown)
library and now uses its parsing infrastructure. On top of that core
functionality, a bunch of improvements have been added:

1. All HTML is now supported natively and dangerouslySetInnerHTML
   is no longer used under any circumstances.

2. Package size is now just ~5kB gzipped, down from ~37kB.

3. Anything rendered via the Markdown component can be overridden now,
   even the wrapper, footnote footer, etc.

I've been working on this release for a while and it took a while
to get all the regexes working to my satisfaction. There will probably
be bugs, so PRs are welcome to add regression tests as needed and
get things tightened up.

Thank you!