Add option to keep all attributes as strings
When data is coming from user supplied HTML, I would expect all attributes to come in as strings.
Instead this library converts boolean strings "true" and "false" to boolean primitives.
My use-case is very heavy on custom components, many of which do string operations on the incoming data.
For example we have one that passes in a a path to get properties off of an object, that path can be the string literal values "true"/"false".
It would be nice to be able to toggle off the feature that converts booleans.