refactor(types): pass types for "reactFor"
I'm not entirely sure what this function is, maybe emitter is a rename of a function that used to be called react?

I was writing what could possibly be a fork (not sure yet) and want to verify I understand the types correctly.

If this is useful, i might find time to resolve the other errors that show up in strict mode too.